feat: Team card moved to another layer
