+++ frontend/src/entities/team/config/index.ts
