+++ frontend/src/entities/team/config/tableHeaders.ts
