+++ frontend/src/entities/team/mock/index.ts
