+++ frontend/src/entities/team/ui/index.ts
@@ -1,1 +1,2 @@
 export { TeamStatsCard } from './TeamStatsCard'
+export { TeamCard } from './TeamCard'
